requirements.txt
accelerate==1.10.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
apache-tvm-ffi==0.1.9
APScheduler==3.11.0
astor==0.8.1
async-timeout==5.0.1
asyncua==1.1.8
attrs==26.1.0
av==16.0.1
backoff==2.2.1
beautifulsoup4==4.14.2
blake3==1.0.8
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
cbor2==5.9.0
certifi==2025.10.5
cffi==2.0.0
chandra-ocr==0.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
coloredlogs==15.0.1
compressed-tensors==0.13.0
contourpy==1.3.2
cryptography==45.0.7
cuda-bindings==12.9.4
cuda-pathfinder==1.5.0
cuda-python==12.9.4
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.20.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
effdet==0.4.1
einops==0.8.1
email-validator==2.3.0
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fairscale==0.4.13
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
filelock==3.17.0
filetype==1.2.0
flashinfer-python==0.6.6
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gguf==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-vision==3.10.2
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
ijson==3.5.0
imageio==2.37.0
importlib_metadata==8.7.1
iniconfig==2.3.0
interegular==0.3.3
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
labml==0.5.3
langdetect==1.0.9
lark==1.2.2
lazy_loader==0.4
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
markdownify==1.1.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.10.0
ml_dtypes==0.5.3
model-hosting-container-standards==0.1.14
mpmath==1.3.0
msgspec==0.20.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numba==0.61.2
numpy==2.1.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.4.2
nvidia-cutlass-dsl-libs-base==4.4.2
nvidia-ml-py==13.595.45
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
openai==2.24.0
openai-clip==1.0.1
openai-harmony==0.0.8
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.5.1
outlines_core==0.2.11
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.9.4
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
poppler-utils==0.1.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.6
psutil==7.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.0.2
pycocotools==2.0.10
pycountry==26.2.16
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==2.0.0
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.12.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.3.0
pypandoc==1.15
pyparsing==3.2.5
pypdf==6.1.2
pypdfium2==4.30.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2025.2.18
python-json-logger==4.1.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==8.0
quack-kernels==0.3.7
RapidFuzz==3.14.1
redis==5.0.8
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-file==3.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.57.0
setproctitle==1.3.7
setuptools==80.10.2
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
six==1.17.0
slack_sdk==3.33.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==3.3.4
starlette==0.52.1
structlog==24.4.0
supervision==0.26.1
supervisor==4.3.0
sympy==1.14.0
tabulate==0.10.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==1.0.20
tldextract==5.1.3
tokenizers==0.22.1
tomli==2.3.0
torch==2.10.0
torch_c_dlpack_ext==0.1.5
torchaudio==2.10.0
torchtext==0.18.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.57.6
triton==3.6.0
twilio==9.3.8
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.217
ultralytics-thop==2.0.17
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.42.0
uvloop==0.22.1
vllm==0.18.1
wait_for2==0.3.2
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
wheel==0.46.3
wrapt==1.17.3
xgrammar==0.1.33
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
zeep==4.3.2
zipp==3.23.0
zxing-cpp==2.2.0
